Add sorted variants of `LiquidityAmounts` functions
Introduced new functions for calculating the amount of liquidity received for a given amount of tokens and price range. These functions, getLiquidityForAmount0Sorted and getLiquidityForAmount1Sorted, work with sorted boundaries. This change simplifies and improves the efficiency of the code by eliminating the need for boundary sort in multiple places. Updated method references in current pool price computations accordingly.