fix: Replace `shr` with `sar` for integer types
This commit replaces `shr` with `sar` in the assembly code of `TickMath.sol`, specifically in commands dealing with integer types. Additionally, it includes an update to the interaction with the variable `tick256` - dropping the `signextend` function and just referencing `tick`. The changes also affect gas snapshots, though it is not the primary purpose of these modifications.